Modularize level1 to pare down the god file
